Update MainActivity.java
Updating to add the method buildHashMapFromFile(). This should read the imagenet.txt file from the assets folder and build a hashmap with keys being indexes and values being class names as strings.

requires txt file of class names to be located in assets folder. May need changing in the future.